+++ src/automata/cyclic.cpp
@@ -38,8 +38,8 @@ std::pair<Board, bool> Cyclic::rewrite(const Board& board) {
     bool change_made = false;
     Board board_copy = board;
 
-    for (int row = 0; row < BOARD_ROWS; row++) {
-        for (int col = 0; col < BOARD_COLS; col++) {
+    for (size_t row = 0; row < BOARD_ROWS; row++) {
+        for (size_t col = 0; col < BOARD_COLS; col++) {
             int next_state = (board[row][col] + 1) % num_states;
 
             int neighbour_count = get_extended_neighbour_count(
